move parser_classes
