+++ tdrs-backend/tdpservice/parsers/parser_classes/tdr_parser.py
@@ -5,10 +5,10 @@ import logging
 from django.conf import settings
 
 from tdpservice.parsers import schema_defs
-from tdpservice.parsers.base_parser import BaseParser
 from tdpservice.parsers.case_consistency_validator import CaseConsistencyValidator
 from tdpservice.parsers.dataclasses import HeaderResult, Position
 from tdpservice.parsers.error_generator import ErrorGeneratorArgs, ErrorGeneratorType
+from tdpservice.parsers.parser_classes.base_parser import BaseParser
 from tdpservice.parsers.schema_defs.utils import ProgramManager
 from tdpservice.parsers.validators import category1
 from tdpservice.parsers.validators.util import value_is_empty
